+++ internal/impl/sql/conn_fields.go
@@ -2,10 +2,34 @@ package sql
 
 import (
 	"database/sql"
+	"net/url"
+	"strings"
 	"time"
 
 	"github.com/benthosdev/benthos/v4/public/service"
 )
+
+var driverField = service.NewStringEnumField("driver", "mysql", "postgres", "clickhouse", "mssql").
+	Description("A database [driver](#drivers) to use.")
+
+var dsnField = service.NewStringField("dsn").
+	Description(`A Data Source Name to identify the target database.
+
+#### Drivers
+
+The following is a list of supported drivers, their placeholder style, and their respective DSN formats:
+
+| Driver | Data Source Name Format |
+|---|---|
+` + "| `clickhouse` | [`clickhouse://[username[:password]@][netloc][:port]/dbname[?param1=value1&...&paramN=valueN]`](https://github.com/ClickHouse/clickhouse-go#dsn) |" + `
+` + "| `mysql` | `[username[:password]@][protocol[(address)]]/dbname[?param1=value1&...&paramN=valueN]` |" + `
+` + "| `postgres` | `postgres://[user[:password]@][netloc][:port][/dbname][?param1=value1&...]` |" + `
+` + "| `mssql` | `sqlserver://[user[:password]@][netloc][:port][?database=dbname&param1=value1&...]` |" + `
+
+Please note that the ` + "`postgres`" + ` driver enforces SSL by default, you can override this with the parameter ` + "`sslmode=disable`" + ` if required.`).
+	Example("clickhouse://username:password@host1:9000,host2:9000/database?dial_timeout=200ms&max_execution_time=60").
+	Example("foouser:foopassword@tcp(localhost:3306)/foodb").
+	Example("postgres://foouser:foopass@localhost:5432/foodb?sslmode=disable")
 
 func connFields() []*service.ConfigField {
 	return []*service.ConfigField{
@@ -26,6 +50,19 @@ func connFields() []*service.ConfigField {
 			Optional().
 			Advanced(),
 	}
+}
+
+func rawQueryField() *service.ConfigField {
+	return service.NewStringField("query").
+		Description("The query to execute. The style of placeholder to use depends on the driver, some drivers require question marks (`?`) whereas others expect incrementing dollar signs (`$1`, `$2`, and so on). The style to use is outlined in this table:" + `
+
+| Driver | Placeholder Style |
+|---|---|
+` + "| `clickhouse` | Dollar sign |" + `
+` + "| `mysql` | Question mark |" + `
+` + "| `postgres` | Dollar sign |" + `
+` + "| `mssql` | Question mark |" + `
+`)
 }
 
 type connSettings struct {
@@ -67,4 +104,31 @@ func connSettingsFromParsed(conf *service.ParsedConfig) (c connSettings, err err
 		}
 	}
 	return
+}
+
+func sqlOpenWithReworks(logger *service.Logger, driver, dsn string) (*sql.DB, error) {
+	if driver == "clickhouse" && strings.HasPrefix("tcp", dsn) {
+		u, err := url.Parse(dsn)
+		if err != nil {
+			return nil, err
+		}
+
+		u.Scheme = "clickhouse"
+
+		u.Path = u.Query().Get("database")
+		u.Query().Del("database")
+
+		if username, password := u.Query().Get("username"), u.Query().Get("password"); username != "" {
+			if password != "" {
+				u.User = url.User(username)
+			} else {
+				u.User = url.UserPassword(username, password)
+			}
+		}
+
+		newDSN := u.String()
+		logger.Warnf("Detected old-style Clickhouse Data Source Name: '%v', replacing with new style: '%v'", dsn, newDSN)
+		dsn = newDSN
+	}
+	return sql.Open(driver, dsn)
 }