+++ internal/impl/sql/processor_sql_raw.go
@@ -21,9 +21,9 @@ func RawProcessorConfig() *service.ConfigSpec {
 If the query fails to execute then the message will remain unchanged and the error can be caught using error handling methods outlined [here](/docs/configuration/error_handling).`).
 		Field(driverField).
 		Field(dsnField).
-		Field(service.NewStringField("query").
-			Description("The query to execute.").
-			Example("INSERT INTO footable (foo, bar, baz) VALUES (?, ?, ?);")).
+		Field(rawQueryField().
+			Example("INSERT INTO footable (foo, bar, baz) VALUES (?, ?, ?);").
+			Example("SELECT * FROM footable WHERE user_id = $1;")).
 		Field(service.NewBoolField("unsafe_dynamic_query").
 			Description("Whether to enable [interpolation functions](/docs/configuration/interpolation/#bloblang-queries) in the query. Great care should be made to ensure your queries are defended against injection attacks.").
 			Advanced().
@@ -96,7 +96,6 @@ type sqlRawProcessor struct {
 	queryStatic string
 	queryDyn    *service.InterpolatedString
 	onlyExec    bool
-	useTxStmt   bool
 
 	argsMapping *bloblang.Executor
 
@@ -143,15 +142,11 @@ func NewSQLRawProcessorFromConfig(conf *service.ParsedConfig, logger *service.Lo
 		}
 	}
 
-	_, useTxStmt := map[string]struct{}{
-		"clickhouse": {},
-	}[driverStr]
-
 	connSettings, err := connSettingsFromParsed(conf)
 	if err != nil {
 		return nil, err
 	}
-	return newSQLRawProcessor(logger, driverStr, dsnStr, queryStatic, queryDyn, onlyExec, useTxStmt, argsMapping, connSettings)
+	return newSQLRawProcessor(logger, driverStr, dsnStr, queryStatic, queryDyn, onlyExec, argsMapping, connSettings)
 }
 
 func newSQLRawProcessor(
@@ -160,7 +155,6 @@ func newSQLRawProcessor(
 	queryStatic string,
 	queryDyn *service.InterpolatedString,
 	onlyExec bool,
-	useTxStmt bool,
 	argsMapping *bloblang.Executor,
 	connSettings connSettings,
 ) (*sqlRawProcessor, error) {
@@ -170,12 +164,11 @@ func newSQLRawProcessor(
 		queryStatic: queryStatic,
 		queryDyn:    queryDyn,
 		onlyExec:    onlyExec,
-		useTxStmt:   useTxStmt,
 		argsMapping: argsMapping,
 	}
 
 	var err error
-	if s.db, err = sql.Open(driverStr, dsnStr); err != nil {
+	if s.db, err = sqlOpenWithReworks(logger, driverStr, dsnStr); err != nil {
 		return nil, err
 	}
 	connSettings.apply(s.db)
@@ -228,28 +221,10 @@ func (s *sqlRawProcessor) ProcessBatch(ctx context.Context, batch service.Messag
 		}
 
 		if s.onlyExec {
-			// TODO: Batch-wide transaction
-			if s.useTxStmt {
-				tx, err := s.db.Begin()
-				if err == nil {
-					_, err = tx.Exec(queryStr, args...)
-				}
-				if err == nil {
-					err = tx.Commit()
-				} else {
-					_ = tx.Rollback()
-				}
-				if err != nil {
-					s.logger.Debugf("Failed to run query: %v", err)
-					msg.SetError(err)
-					continue
-				}
-			} else {
-				if _, err := s.db.ExecContext(ctx, queryStr, args...); err != nil {
-					s.logger.Debugf("Failed to run query: %v", err)
-					msg.SetError(err)
-					continue
-				}
+			if _, err := s.db.ExecContext(ctx, queryStr, args...); err != nil {
+				s.logger.Debugf("Failed to run query: %v", err)
+				msg.SetError(err)
+				continue
 			}
 		} else {
 			rows, err := s.db.QueryContext(ctx, queryStr, args...)