+++ internal/impl/sql/resources/clickhouse_init.sql
@@ -1,0 +1,5 @@
+create table test (
+  foo String,
+  bar Int64,
+  baz String
+) engine=Memory